+++ sdk/eventhubs/azure_messaging_eventhubs/tests/eventhubs_producer.rs
@@ -81,6 +81,8 @@ async fn get_properties(ctx: TestContext) -> Result<(), Box<dyn Error>> {
     info!("Properties: {:?}", properties);
     assert_eq!(properties.name, eventhub);
 
+    client.close().await?;
+
     Ok(())
 }
 
@@ -137,6 +139,8 @@ async fn get_partition_properties(ctx: TestContext) -> Result<(), Box<dyn Error>
         let amqp_error = err.source().unwrap().downcast_ref::<AmqpError>().unwrap();
         info!("AMQP error: {:?}", amqp_error);
     }
+
+    client.close().await?;
 
     Ok(())
 }
@@ -204,6 +208,8 @@ async fn send_eventdata(ctx: TestContext) -> Result<(), Box<dyn Error>> {
     // Simple send.
     assert!(client.send_event("Hello, Event Hub!", None).await.is_ok());
 
+    client.close().await?;
+
     Ok(())
 }
 
@@ -249,6 +255,8 @@ async fn send_message(ctx: TestContext) -> Result<(), Box<dyn Error>> {
     // Simple send.
     assert!(client.send_event("Hello, Event Hub!", None).await.is_ok());
 
+    client.close().await?;
+
     Ok(())
 }
 
@@ -268,6 +276,7 @@ async fn test_create_batch(ctx: TestContext) -> Result<(), Box<dyn Error>> {
         let batch = client.create_batch(None).await?;
         assert_eq!(batch.len(), 0);
     }
+    client.close().await?;
 
     Ok(())
 }
@@ -316,6 +325,8 @@ async fn test_create_and_send_batch(ctx: TestContext) -> Result<(), Box<dyn Erro
         let res = client.send_batch(batch, None).await;
         assert!(res.is_ok());
     }
+
+    client.close().await?;
 
     Ok(())
 }
@@ -388,6 +399,8 @@ async fn test_add_amqp_messages_to_batch(
 
     client.send_batch(batch, None).await?;
 
+    client.close().await?;
+
     Ok(())
 }
 
@@ -446,6 +459,8 @@ async fn test_overload_batch(ctx: TestContext) -> Result<(), Box<dyn Error>> {
         }
         assert!(result.is_ok());
     }
+
+    client.close().await?;
 
     Ok(())
 }