+++ sdk/storage/azure_storage_queue/src/clients/mod.rs
@@ -1,0 +1,10 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+mod queue_client;
+pub use queue_client::QueueClient;
+
+mod queue_service_client;
+pub use queue_service_client::QueueServiceClient;
+
+pub use crate::generated::clients::{QueueClientOptions, QueueServiceClientOptions};