+++ sdk/typespec/typespec_client_core/src/http/response.rs
@@ -4,13 +4,14 @@
 //! HTTP responses.
 
 use crate::{
-    http::{headers::Headers, DeserializeWith, Format, JsonFormat, RawResponse, StatusCode},
+    http::{headers::Headers, DeserializeWith, Format, JsonFormat, StatusCode},
     Bytes,
 };
 use futures::{Stream, StreamExt};
 use serde::de::DeserializeOwned;
 use std::{fmt, marker::PhantomData, pin::Pin};
 use typespec::error::{ErrorKind, ResultExt};
+pub use typespec::http::response::*;
 
 /// A pinned stream of bytes that can be sent as a response body.
 #[cfg(not(target_arch = "wasm32"))]
@@ -411,7 +412,7 @@ mod tests {
                 .iter()
                 .any(|(k, v)| k.as_str() == "x-ms-error" && v.as_str() == "BadParameter"));
 
-            let err: ErrorResponse = err.json().await.expect("convert to ErrorResponse");
+            let err: ErrorResponse = err.json().expect("convert to ErrorResponse");
             assert_eq!(err.code, Some("BadParameter".into()));
             assert_eq!(err.message, Some("Bad parameter".into()));
         }